requirements.txt
torch>=2.0.0
numpy>=1.21.0
gymnasium>=0.28.0
pygame>=2.5.0
matplotlib>=3.5.0
pyyaml>=6.0
tqdm>=4.65.0
tensorboard>=2.12.0
opencv-python>=4.8.0
pillow>=9.5.0
